added comments to code and cleaned up
