+++ src/modifiers/mod_func_door_setup.c
@@ -34,54 +34,21 @@ void	mod_gen_dr(char *content, int index, t_world *world, t_map *map)
 	int	i;
 	char	**dr;
 	char	*texture;
-
-	i = -1;
-//	map->mods[index].type = DOOR;
-	dr = ft_split(content, ',');
-	while (dr[0][++i] && i < NAME_SIZE)
-		world->ent->doors[dr_count].name[i] = dr[0][i];
-	map->mods[index].content = ft_strdup(dr[1]);
-	world->ent->doors[dr_count].pos.x = ft_atoi(dr[2]);
-	world->ent->doors[dr_count].pos.y = ft_atoi(dr[3]);
-	world->ent->doors[dr_count].speed = ft_atoi(dr[4]);
-	world->ent->doors[dr_count].closed = (dr[5][0] == 'C') ? true : false;
-	world->ent->doors[dr_count].locked = (dr[6][0] == 'L') ? true : false;
-
-
-
-	map->tiles[ft_atoi(dr[3]) * map->width + ft_atoi(dr[2]) -1].type = DOOR;
-	map->tiles[ft_atoi(dr[3]) * map->width + ft_atoi(dr[2]) -1].vis = 1;
-	map->tiles[ft_atoi(dr[3]) * map->width + ft_atoi(dr[2]) -1].tex = 6;
-
-//new entity_2 stuff
-	i = -1;
 	int	x;
 	int	y;
 
+	i = -1;
+	dr = ft_split(content, ',');
+	map->tiles[ft_atoi(dr[3]) * map->width + ft_atoi(dr[2]) -1].type = DOOR;
+	map->tiles[ft_atoi(dr[3]) * map->width + ft_atoi(dr[2]) -1].vis = 1;
+	map->tiles[ft_atoi(dr[3]) * map->width + ft_atoi(dr[2]) -1].tex = 6;
 	x = ft_atoi(dr[2]) - 1;
 	y = ft_atoi(dr[3]);
 	while (dr[0][++i] && i < NAME_SIZE)
 		world->ent_2[world->ent_count].name[i] = dr[0][i];
 	ft_strcpy(world->ent_2[world->ent_count].ui_display_1, "PRESS 'E' TO INTERACT");
 	ft_strcpy(world->ent_2[world->ent_count].ui_display_2, "LOCKED");
-	world->ent_2[world->ent_count].pos[0].x = x;
-	world->ent_2[world->ent_count].pos[0].y = y;
-	world->ent_2[world->ent_count].pos[1].x = x + 1;
-	world->ent_2[world->ent_count].pos[1].y = y;
-	world->ent_2[world->ent_count].pos[2].x = x + 1;
-	world->ent_2[world->ent_count].pos[2].y = y + 1;
-	world->ent_2[world->ent_count].pos[3].x = x;
-	world->ent_2[world->ent_count].pos[3].y = y + 1;
-	world->ent_2[world->ent_count].pos[4].x = x - 1;
-	world->ent_2[world->ent_count].pos[4].y = y + 1;
-	world->ent_2[world->ent_count].pos[5].x = x - 1;
-	world->ent_2[world->ent_count].pos[5].y = y;
-	world->ent_2[world->ent_count].pos[6].x = x - 1;
-	world->ent_2[world->ent_count].pos[6].y = y - 1;
-	world->ent_2[world->ent_count].pos[7].x = x;
-	world->ent_2[world->ent_count].pos[7].y = y - 1;
-	world->ent_2[world->ent_count].pos[8].x = x + 1;
-	world->ent_2[world->ent_count].pos[8].y = y - 1;
+	get_pos_and_surrounding_tiles(world, x, y);
 	world->ent_2[world->ent_count].ref_tile = &map->tiles[y * (map->width + x) - 1];
 	world->ent_2[world->ent_count].target = NULL;
 	world->ent_2[world->ent_count].speed = ft_atoi(dr[4]);
@@ -122,3 +89,19 @@ void	mod_gen_ke(char *content, int index, t_world *world, t_map *map)
 	world->ent->keys[ke_count].collected = (ke[6][0] == 'C') ? true : false;
 	free_str_array(ke);
 }
+
+
+
+
+
+/*
+	while (dr[0][++i] && i < NAME_SIZE)
+		world->ent->doors[dr_count].name[i] = dr[0][i];
+	map->mods[index].content = ft_strdup(dr[1]);
+
+	world->ent->doors[dr_count].pos.x = ft_atoi(dr[2]);
+	world->ent->doors[dr_count].pos.y = ft_atoi(dr[3]);
+	world->ent->doors[dr_count].speed = ft_atoi(dr[4]);
+	world->ent->doors[dr_count].closed = (dr[5][0] == 'C') ? true : false;
+	world->ent->doors[dr_count].locked = (dr[6][0] == 'L') ? true : false;
+*/
