+++ src/vector/vector2_math_extra.c
@@ -50,3 +50,8 @@ t_vec2	v2diff(t_vec2 f, t_vec2 s)
 		vec.y = s.y - f.y;
 	return (vec);
 }
+
+double v2dot(t_vec2 a, t_vec2 b)
+{
+  return (a.x * b.x + a.y * b.y);
+}
